requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
redis>=4.5.0
rq>=1.15.0
pymongo>=4.3.0
aiohttp>=3.8.0
asyncio>=3.4.3
requests>=2.28.0
python-dotenv==1.0.1
